[aws-datastore] Move AppSync type enumerations out of core
The Core Model schema includes an enumeration of some various
`JavaFieldType`. This will remain in the Core Models package.

The AWS DataStore implementation maps between these types, and the
various `SqliteDataType`, and `AWSAppSyncScalarType`. These types, and
the business of mapping them, belongs as a detail of the AWS DataStore
plugin.

As consequences of this change:

1. AppSync business details are not part of Amplify's core
   specification.
2. If the CLI shall generate references to any of these types, it will
   bind the implementation specifically for use of the AWS DataStore
   Plugin.
