+++ aws-datastore/src/main/java/com/amplifyframework/datastore/appsync/AWSAppSyncScalarType.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package com.amplifyframework.core.model.types;
+package com.amplifyframework.datastore.appsync;
 
 import androidx.annotation.NonNull;
