+++ aws-datastore/src/main/java/com/amplifyframework/datastore/storage/sqlite/adapter/SQLiteColumn.java
@@ -16,7 +16,7 @@
 package com.amplifyframework.datastore.storage.sqlite.adapter;
 
 import com.amplifyframework.core.model.PrimaryKey;
-import com.amplifyframework.core.model.types.SqliteDataType;
+import com.amplifyframework.datastore.storage.sqlite.SqliteDataType;
 
 /**
  * Adapts a {@link com.amplifyframework.core.model.ModelField}
