+++ aws-datastore/src/main/java/com/amplifyframework/datastore/storage/sqlite/adapter/SQLiteTable.java
@@ -24,8 +24,8 @@ import com.amplifyframework.core.model.ModelField;
 import com.amplifyframework.core.model.ModelSchema;
 import com.amplifyframework.core.model.PrimaryKey;
 import com.amplifyframework.core.model.types.JavaFieldType;
-import com.amplifyframework.core.model.types.SqliteDataType;
-import com.amplifyframework.core.model.types.internal.TypeConverter;
+import com.amplifyframework.datastore.storage.sqlite.SqliteDataType;
+import com.amplifyframework.datastore.storage.sqlite.TypeConverter;
 import com.amplifyframework.util.Immutable;
 
 import java.util.Collections;
